fix(scoring): widen score distribution and add decimal precision
Scores were clustered at 8/10 due to: example JSON anchoring (7-9),
missing guidance for 4-6 range, integer rounding compression, and
"5 = indistinguishable" setting a high floor.

- Rewrite validator rubric with full 1-10 calibration (every level
  defined) including anti-patterns that warrant low scores (3-5)
- Change example JSON from {8,7,9,8,7} to {6,5,7,8,6} to anchor
  moderate range as normal
- Add "Be rigorous" instruction: reserve 8+ for demonstrable quality
  transformations
- Change overall_score from Integer to Float (1 decimal place) to
  preserve differentiation (7.2 vs 8.1 instead of both rounding to 8)
- Update compute_overall_score return type: int → float, round(raw) →
  round(raw, 1)
- Change LOW_SCORE_THRESHOLD from int 5 to float 5.0 for type clarity

SQLite dynamic typing handles Integer→Float transparently; no
migration required.

+++ backend/app/models/optimization.py
@@ -3,7 +3,7 @@ import logging
 import uuid
 from datetime import datetime, timezone
 
-from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Index, Integer, String, Text
+from sqlalchemy import Boolean, Column, DateTime, Float, ForeignKey, Index, Integer, String, Text
 
 from app.database import Base
 
@@ -48,7 +48,7 @@ class Optimization(Base):
     structure_score = Column(Integer, nullable=True)
     faithfulness_score = Column(Integer, nullable=True)
     conciseness_score = Column(Integer, nullable=True)
-    overall_score = Column(Integer, nullable=True)
+    overall_score = Column(Float, nullable=True)
 
     # Validation verdict
     is_improvement = Column(Boolean, nullable=True)
